Fix reindex feature on error.
Added a new dialog so that the user knows the wallet will be restarted
and should wait some seconds.

Had to change the approach and wait some seconds with daemon running
with reindex before stopping it and also wait some seconds to check if
it stopped properly before restart the wallet.